+++ zhixuewang/teacher/__init__.py
@@ -1,6 +1,7 @@
 from zhixuewang.teacher.teacher import TeacherAccount
+from zhixuewang.teacher import workbook
 
 __all__ = [
     "TeacherAccount",
-
+    "workbook"
 ]
